+++ Changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased Version 1.9.0 - 2026-XX-XX](https://github.com/rgadon107/custom-functionality/pull/14)
+### Added
+
+### Changed
+
+### Fixed
+
+### Corrections
+
 ## [Version 1.8.0 - 2026-05-10](https://github.com/rgadon107/custom-functionality/pull/13)
 ### Added
 - `/includes/patterns/`: Added directory to manage logic to register design patterns with WordPress.